refactor(ui): вынесено toolbar state для SidebarView
subtitle и syncing-derived state sidebar toolbar перенесены в SidebarToolbarState и SidebarSubtitleFormatter, чтобы производная логика больше не жила внутри SidebarView
